EDIT: login page enter event, main page comment enter event, main page component
