+++ src/pages/minseok/Main/CommentList.js
@@ -5,6 +5,8 @@ const CommentList = ({ comment }) => {
     <li className="postingCommentList">
       <span className="userCommentId">김펭귄</span>
       <span className="userComment">{comment}</span>
+      <button className="userCommentLikeHeart">하트</button>
+      <button className="userCommentRemove">삭제</button>
     </li>
   );
 };
