+++ src/pages/minseok/Main/main.scss
@@ -172,19 +172,29 @@ html {
           background-color: #ffffff;
           font-size: 1.3rem;
 
+          .postingCommentList {
+            .userCommentId {
+              font-weight: 700;
+              font-size: 1.6rem;
+            }
+
+            .userComment {
+              margin-top: 0.5rem;
+              margin-left: 1rem;
+              font-size: 1.3rem;
+            }
+
+            .userCommentLikeHeart {
+              margin-left: 1rem;
+            }
+
+            .userCommentRemove {
+              margin-left: 1rem;
+            }
+          }
+
           .RepresentativeUser {
             margin-bottom: 1rem;
-          }
-
-          .userCommentId {
-            font-weight: 700;
-            font-size: 1.6rem;
-          }
-
-          .userComment {
-            margin-top: 0.5rem;
-            margin-left: 1rem;
-            font-size: 1.3rem;
           }
         }
